Fixed Mounts Fixed Dash and Speed Infusion (No idea if it crashes the map) Some ring stuff
git-svn-id: https://svn6.assembla.com/svn/MoopleDEV@70 59296cf3-ab24-4890-a14e-c0dede3ece5d

+++ src/server/MaplePlayerShop.java
@@ -127,7 +127,7 @@ public class MaplePlayerShop extends AbstractMapleMapObject {
                 return;
             }
             synchronized (c.getPlayer()) {
-                if (c.getPlayer().getMeso() >= pItem.getPrice() * quantity) {
+                if (c.getPlayer().getMeso() >= (long) pItem.getPrice() * quantity) {
                     if (MapleInventoryManipulator.addFromDrop(c, newItem, false)) {
                         c.getPlayer().gainMeso(-pItem.getPrice() * quantity, true);
                         owner.gainMeso(pItem.getPrice() * quantity, true);
